+++ examples/example_wall_slide_and_jump.rs
@@ -31,6 +31,7 @@ fn main() {
 enum ControlScheme {
     Jump(TnuaBuiltinJump),
     WallSlide(TnuaBuiltinWallSlide, Entity),
+    #[scheme(same_trigger(Jump))]
     WallJump(TnuaBuiltinJump),
 }