+++ macros/src/scheme_derive/gen_action_discriminant.rs
@@ -1,3 +1,6 @@
+use std::collections::HashMap;
+
+use atterate::StaticError;
 use proc_macro2::TokenStream;
 use quote::quote;
 
@@ -11,10 +14,36 @@ pub fn generate_action_discriminant(parsed: &ParsedScheme) -> syn::Result<TokenS
         ..
     } = parsed;
     let command_names = commands.iter().map(|c| c.command_name).collect::<Vec<_>>();
-    let variant_indices = commands
+
+    let mut same_trigger_targets: HashMap<&syn::Ident, Option<usize>> = commands
         .iter()
-        .enumerate()
-        .map(|(i, _)| syn::Index::from(i));
+        .filter_map(|c| Some((c.same_trigger.as_ref()?, None)))
+        .collect();
+    if !same_trigger_targets.is_empty() {
+        for (i, cmd) in commands
+            .iter()
+            .filter(|c| c.same_trigger.is_none())
+            .enumerate()
+        {
+            if let Some(target) = same_trigger_targets.get_mut(cmd.command_name) {
+                *target = Some(i);
+            }
+        }
+    }
+    let same_trigger_targets: HashMap<&syn::Ident, usize> = same_trigger_targets
+        .into_iter()
+        .map(|(k, v)| Ok((k, v.ok_or(StaticError::Spanned(k, "Command variant does not exist in scheme, or exists but has a `same_trigger` of its own"))?)))
+        .collect::<Result<_, syn::Error>>()?;
+
+    let mut next_index = 0..;
+    let variant_indices = commands.iter().map(|command| {
+        let variant_index = if let Some(same_as) = command.same_trigger.as_ref() {
+            same_trigger_targets[same_as]
+        } else {
+            next_index.next().expect("endless range should not end")
+        };
+        syn::Index::from(variant_index)
+    });
 
     let (serde_derives, serde_attr) = parsed.gen_serde_clauses_always();
 
@@ -33,8 +62,8 @@ pub fn generate_action_discriminant(parsed: &ParsedScheme) -> syn::Result<TokenS
                     #(
                         Self::#command_names => #variant_indices,
                     )*
-                    #[allow(unreachable_patterns)]
-                    _ => unreachable!(),
+                        #[allow(unreachable_patterns)]
+                        _ => unreachable!(),
                 }
             }
         }