+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 NOTE: Subcrates have their own changelogs: [bevy-tnua-physics-integration-layer](physics-integration-layer/CHANGELOG.md), [bevy-tnua-rapier](rapier3d/CHANGELOG.md), [bevy-tnua-avian](avian3d/CHANGELOG.md).
 
 ## [Unreleased]
+### Fixed
+- [**BREAKING**] `#[scheme]` attribute on `TnuaScheme` variants now generate
+  errors when passed invalid arguments.
+
+### Added
+- `#[scheme(same_trigger = ...)]` parameter for the `TnuaScheme` derive macro.
 
 ## 0.30.0 - 2026-01-31
 ### Added